Ground Item Rework
This completely overhauls the overworld and hidden items in PokeScape so that these scale better with progression (e.g. you do not get bread after a few gyms) as well as generally increase the quantity of pouches being picked up to reward and encourage catching.

A full list of items will also be available on the spreadsheet moving forward.