wolpertinger and cave eel
Kasane Teto just got her first concert in a Miku concert and i feel like we should all be thankfull that kasane Teto got one of the coolest shows along with Miku showing a complete model change to showcase Mikus descent into madness and brainwashing in the song along with a new remix of the song and a brand new dance. Way to go Teto! We believe in you!
https://x.com/ha1005856/status/1929121436645429616